+++ app/Http/Controllers/Back/Assistants/PublishController.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Thinktomorrow\Chief\App\Http\Controllers\Back;
+namespace Thinktomorrow\Chief\App\Http\Controllers\Back\Assistants;
 
 use Illuminate\Http\Request;
 use Thinktomorrow\Chief\Management\Managers;
 use Thinktomorrow\Chief\App\Http\Controllers\Controller;
-use Thinktomorrow\Chief\Management\NotAllowedManagerRoute;
 
 class PublishController extends Controller
 {
@@ -16,19 +15,15 @@ class PublishController extends Controller
     {
         $this->managers = $managers;
     }
-    
+
     public function publish(Request $request, $key, $id)
     {
         $manager = $this->managers->findByKey($key, $id);
 
-        $manager->guard('publish');
+        $manager->assistant('publish')
+                ->guard('publish')
+                ->publish();
 
-        if (!isManagerThatPublishes($manager)) {
-            throw NotAllowedManagerRoute::notAllowedVerb('publish', $manager);
-        }
-
-        $manager->publish();
-        
         return redirect()->back()->with('messages.success', $manager->details()->title .' is gepubliceerd. <a href="' . $manager->previewUrl() . '" target="_blank">Bekijk de pagina online</a>.');
     }
 
@@ -36,13 +31,9 @@ class PublishController extends Controller
     {
         $manager = $this->managers->findByKey($key, $id);
 
-        $manager->guard('draft');
-
-        if (!isManagerThatPublishes($manager)) {
-            throw NotAllowedManagerRoute::notAllowedVerb('draft', $manager);
-        }
-
-        $manager->draft();
+        $manager->assistant('publish')
+                ->guard('draft')
+                ->draft();
 
         return redirect()->back();
     }