+++ resources/views/back/managers/_partials/publish-option.blade.php
@@ -1,18 +1,18 @@
-@if(isManagerThatPublishes($manager))
-    @if($manager->isPublished())
+@if($manager->isAssistedBy('publish'))
+    @if($manager->assistant('publish')->isPublished())
 
         <a data-submit-form="draftForm-{{ $manager->details()->id }}" class="block squished-s text-warning --link-with-bg">Haal offline</a>
 
-        <form class="--hidden" id="draftForm-{{ $manager->details()->id }}" action="{{ $manager->route('draft') }}" method="POST">
+        <form class="--hidden" id="draftForm-{{ $manager->details()->id }}" action="{{ $manager->assistant('publish')->route('draft') }}" method="POST">
             {{ csrf_field() }}
             <button type="submit">Unpublish</button>
         </form>
 
-    @elseif($manager->isDraft())
+    @elseif($manager->assistant('publish')->isDraft())
 
         <a data-submit-form="publishForm-{{ $manager->details()->id }}" class="block squished-s --link-with-bg">Zet online</a>
 
-        <form class="--hidden" id="publishForm-{{ $manager->details()->id }}" action="{{ $manager->route('publish') }}" method="POST">
+        <form class="--hidden" id="publishForm-{{ $manager->details()->id }}" action="{{ $manager->assistant('publish')->route('publish') }}" method="POST">
             {{ csrf_field() }}
             <button type="submit">Publish</button>
         </form>