+++ src/Common/helpers.php
@@ -104,14 +104,6 @@ if (! function_exists('contract')) {
     }
 }
 
-if (! function_exists('isManagerThatPublishes')) {
-    function isManagerThatPublishes($class)
-    {
-        return contract($class, \Thinktomorrow\Chief\Management\ManagerThatPublishes::class);
-    }
-}
-
-
 /**
  * --------------------------------------------------------------------------
  * Helper: Teaser