+++ src/Fields/Types/Field.php
@@ -82,9 +82,9 @@ class Field
         return false;
     }
 
-    public static function translateValue($value, $locale)
+    public static function translateValue($value, $locale = null)
     {
-        if (!is_array($value)) {
+        if(!$locale || !is_array($value)) {
             return $value;
         }